+++ CHANGELOG.md
@@ -7,12 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.0.1](https://github.com/Akaizoku/PSAYX/compare/1.0.0...main) - Unreleased
 
-Robustness improvements
+Robustness improvements and bug fixes
 
 ### Changed
 
 The following functions have been updated:
 
+- Get-Utility
 - Start-Workflow
 
 ## [1.0.0](https://github.com/Akaizoku/PSAYX/releases/tag/1.0.0) - 2021-09-20